CMake: Only run the tests discovered by `gtest_discover_tests`
There is an open [CMake issue](https://gitlab.kitware.com/cmake/cmake/-/issues/20212)
to not run tests in sub-projects, but it has seen no activity for a while, so
we work around it by only running tests whose names start with an uppercase
letter. This works, as `gtest_discover_tests` enumerates only the tests in the
local project and the Google style guide mandates test names begin with an
uppercase letter. \
This ignores any other tests that CTest may find.

PiperOrigin-RevId: 479514038
Change-Id: I4c705f6b8f7f51fb683d94c9a6e81afa34c1381e